Progress Update...
Implement most of the functionality for the first part of the java portion of the assingment. I've written a part of the second part so that I can test the readibility of heap file. I've commented that out but I can confirm that it works. Running the code on the full dataset takes several seconds, but I have confirmed that it works fine.