Progress Update...
Full functionality for both parts of the java aspect of assignment. The heapfile is able to completely load the files as long as they are in the specified format (hardcoded the number of bytes per a field). Loading anything besides a file in the same layout with the same approximate field size will kill the program. DBQuery will look for a heapfile with by matching the page size so there can't be a heapfile with the same page size (i.e 4096) because there can only be one file called heap.4096. Asides from that, there is no commenting yet for either file. I will go and do that now.